requirements.txt
Flask
gunicorn
discord_webhook
Flask-Limiter
Flask-Mobility
discord.py
python-dotenv
schedule
beautifulsoup4